// src/types/type.ts
export type TransactionStatus = "pending" | "success" | "failed";
export type SortDirection = "asc" | "desc";
export type SortColumn =
  | "name"
  | "date"
  | "invoiceId"
  | "amount"
  | "status"
  | "";
// ...
export interface Transaction {
  id: string;
  name: string;
  category: string;
  date: string;
  time: string;
  invoiceId: string;
  amount: string;
  status: TransactionStatus;
  imageUrl?: string;
  avatarInitial?: string;
  avatarColor?: string;
  recipient?: string;
  recipientEmail?: string;
  accountNumber?: string;
  transactionFee?: string;
  spendingHistory?: { month: string; value: number }[];
}